+++ src/code_1b40_len_20b0.c
@@ -1,31 +1,38 @@
 #include "common.h"
 
-extern Gfx D_80074210[];
-extern Gfx D_80074230[];
-
 s8 D_80074020 = 1;
 s8 D_80074021 = 5;
 
 GameStatus gGameStatus = {0};
-
 GameStatus* gGameStatusPtr[1] = { &gGameStatus };
-
 s16 D_800741A0 = 0;
-
 s16 D_800741A2 = 0;
-
 s32 D_800741A4 = 0;
-
-s32 D_800741A8[] = { 0x00010000, 0x00000000, 0x00000001, 0x00000000, 0x00000000, 0x00010000, 0x00000000, 0x00000001, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, };
-
+s32 D_800741A8[] = { 0x00010000, 0x00000000, 0x00000001, 0x00000000, 0x00000000, 0x00010000, 0x00000000, 0x00000001,
+                     0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000, 0x00000000,
+                     0x00000000, 0x00000000, };
 u16 gMatrixListPos = 0;
-
 u16 D_800741F2 = 0;
-
 u16 gCurrentDisplayContextIndex = 0;
+s32 D_800741F8 = 0;
+s32 D_800741FC = 0;
+s32 D_80074200[] = { 0x028001E0, 0x01FF0000, 0x028001E0, 0x01FF0000 };
 
-s32 D_800741F8[] = { 0x00000000, 0x00000000, 0x028001E0, 0x01FF0000, 0x028001E0, 0x01FF0000, 0xE200001C, 0x0F0A4000, 0xFCFFFFFF, 0xFFFE793C, 0xE3001801, 0x00000040, 0xDF000000, 0x00000000, 0xDC080008, 0x80074200, 0xD9000000, 0x00000000, 0xD9FFFFFF, 0x00200405, 0xD7000000, 0x00000000, 0xDF000000, 0x00000000, 0x00000000, 0x00000000, };
+Gfx D_80074210[] = {
+    gsDPSetRenderMode(G_RM_OPA_SURF, G_RM_OPA_SURF2),
+    gsDPSetCombineMode(G_CC_SHADE, G_CC_SHADE),
+    gsDPSetColorDither(G_CD_BAYER),
+    gsSPEndDisplayList(),
+};
 
+Gfx D_80074230[] = {
+    gsSPViewport(&D_80074200),
+    gsSPClearGeometryMode(G_ZBUFFER | G_SHADE | G_CULL_BOTH | G_FOG | G_LIGHTING | G_TEXTURE_GEN |
+                          G_TEXTURE_GEN_LINEAR | G_LOD | G_SHADING_SMOOTH | G_CLIPPING | 0x0040F9FA),
+    gsSPSetGeometryMode(G_ZBUFFER | G_SHADE | G_CULL_BACK | G_SHADING_SMOOTH),
+    gsSPTexture(0, 0, 0, G_TX_RENDERTILE, G_OFF),
+    gsSPEndDisplayList(),
+};
 
 INCLUDE_ASM(void, "code_1b40_len_20b0", step_game_loop);
 
@@ -33,23 +40,24 @@ INCLUDE_ASM(s32, "code_1b40_len_20b0", gfx_task_background);
 
 INCLUDE_ASM(s32, "code_1b40_len_20b0", gfx_draw_frame);
 
-#ifdef NON_MATCHING
 void load_engine_data(void) {
-    GameStatus* phi_s0;
+    GameStatus** gameStatusPtrTemp = gGameStatusPtr;
+    GameStatus** gameStatusPtrTemp2 = gGameStatusPtr;
+    GameStatus* gameStatus3;
     s32 i;
 
-    dma_copy(code_code_FEE30_ROM_START, code_code_FEE30_ROM_END, code_code_FEE30_VRAM);
-    dma_copy(code_code_759B0_ROM_START, code_code_759B0_ROM_END, code_code_759B0_VRAM);
-    dma_copy(code_code_E79B0_ROM_START, code_code_E79B0_ROM_END, code_code_E79B0_VRAM);
-    dma_copy(code_code_102610_ROM_START, code_code_102610_ROM_END, code_code_102610_VRAM);
-    dma_copy(code_code_A5DD0_ROM_START, code_code_A5DD0_ROM_END, code_code_A5DD0_VRAM);
-    dma_copy(D_0010CC10, D_0010CC10_end, D_802EB3D0); // new segment required to fix this
+    dma_copy(&code_code_FEE30_ROM_START, &code_code_FEE30_ROM_END, &code_code_FEE30_VRAM);
+    dma_copy(&code_code_759B0_ROM_START, &code_code_759B0_ROM_END, &code_code_759B0_VRAM);
+    dma_copy(&code_code_E79B0_ROM_START, &code_code_E79B0_ROM_END, &code_code_E79B0_VRAM);
+    dma_copy(&code_code_102610_ROM_START, &code_code_102610_ROM_END, &code_code_102610_VRAM);
+    dma_copy(&code_code_A5DD0_ROM_START, &code_code_A5DD0_ROM_END, &code_code_A5DD0_VRAM);
+    dma_copy(&code_code_10CC10_ROM_START, &code_code_10CC10_ROM_END, &code_code_10CC10_VRAM);
+
     D_8009A650[0] = 0;
-    phi_s0 = *gGameStatusPtr;
     (*gGameStatusPtr)->unk_79 = 0;
+    (*gameStatusPtrTemp2)->enableBackground = 0;
     (*gGameStatusPtr)->musicEnabled = 1;
-    phi_s0->unk_148 = 0;
-    (*gGameStatusPtr)->unk_7C = 1;
+    (*gameStatusPtrTemp)->unk_7C = 1;
     (*gGameStatusPtr)->unk_A8 = -1;
     (*gGameStatusPtr)->unk_AA = 0;
     (*gGameStatusPtr)->unk_81 = 0;
@@ -58,12 +66,12 @@ void load_engine_data(void) {
     D_8009A5D8 = 0;
     (*gGameStatusPtr)->unk_75 = (*gGameStatusPtr)->unk_13C = 0;
     D_80074021 = 5;
-    (*gGameStatusPtr)->unk_168 = 0;
+    (*gGameStatusPtr)->saveCount = 0;
     fio_init_flash();
     func_80028838();
     general_heap_create();
     func_8011D890();
-    func_801230B8();
+    clear_dynamic_entity_list();
     clear_script_list();
     create_cameras_a();
     clear_player_status();
@@ -76,10 +84,10 @@ void load_engine_data(void) {
     clear_character_set();
     clear_printers();
     func_80112B98();
-    func_800384BC();
+    clear_npcs();
     func_80141100();
     clear_trigger_data();
-    func_80110CB8(0);
+    clear_entity_data(0);
     clear_player_data();
     func_8003E338();
     clear_transition_stencil();
@@ -90,21 +98,18 @@ void load_engine_data(void) {
     reset_ambient_sounds();
     func_80149618();
     clear_windows();
-    curtains_init();
+    initialize_curtains();
     poll_rumble();
 
-    phi_s0 = *gGameStatusPtr;
+    gameStatus3 = gGameStatusPtr[0];
     for (i = 0; i < 4; i++) {
-        phi_s0->unk_50[i] = 3;
-        phi_s0->unk_48[i] = 0xC;
+        gameStatus3->unk_50[i] = 3;
+        gameStatus3->unk_48[i] = 0xC;
     }
 
     D_8009A650[0] |= 8;
     set_game_mode(0);
 }
-#else
-INCLUDE_ASM(s32, "code_1b40_len_20b0", load_engine_data);
-#endif
 
 //weird ordering issue
 #ifdef NON_MATCHING
@@ -149,19 +154,16 @@ s32 func_80027190(void) {
 
 #ifdef NON_MATCHING
 void gfx_init_state(void) {
+    Gfx* temp;
+
     gSPSegment(gMasterGfxPos++, 0x00, 0x0);
     gSPDisplayList(gMasterGfxPos++, OS_K0_TO_PHYSICAL(&D_80074230));
+    temp = gMasterGfxPos++;
     gSPDisplayList(gMasterGfxPos++, OS_K0_TO_PHYSICAL(&D_80074210));
 }
 #else
 INCLUDE_ASM(s32, "code_1b40_len_20b0", gfx_init_state);
 #endif
-
-// void gfx_init_state(void) {
-//     gSPSegment(gMasterGfxPos++, 0x00, 0x00000000);
-//     gSPDisplayList(gMasterGfxPos++, D_00074230);
-//     gSPDisplayList(gMasterGfxPos++, D_00074210);
-// }
 
 INCLUDE_ASM(s32, "code_1b40_len_20b0", func_800271FC);
