+++ ver/us/splat.yaml
@@ -200,7 +200,7 @@ segments:
     - [0x457c0, c, os/code_457c0_len_270]
     - [0x45a30, c, os/code_45a30_len_20]
     - [0x45a50, c, os/code_45a50_len_3a0]
-    - [0x45df0, hasm, os/exceptasm]
+    - [0x45df0, hasm, os/exceptasm, skip]
     - [0x46760, hasm, os/__osDisableInt]
     - [0x467D0, hasm, os/__osRestoreInt]
     - [0x467f0, c, os/code_467f0_len_90]
@@ -427,6 +427,7 @@ segments:
     - [0x107830, c, code_107830_len_e70]
     - [0x1086a0, c, code_1086a0_len_fc0]
     - [0x109660, c, code_109660_len_1270]
+    - [0x10A230, c, code_10A230]
     - [0x10A8D0, c, code_10A8D0]
     - [0x10A9F0, data]
     #- [0x10A9F0, i4,      shadow/square, 16, 16]
@@ -443,7 +444,13 @@ segments:
     - [0x10CB30, .rodata, code_107830_len_e70]
     - [0x10CB70, .rodata, code_1086a0_len_fc0]
     - [0x10CBD0, .rodata, code_109660_len_1270]
-  - [0x10CC00, bin]
+    - [0x10CC00, .rodata, code_10A230]
+  - type: code
+    start: 0x10CC10
+    vram: 0x802EB3D0
+    subsections:
+    - [0x10CC10, bin]
+  - [0x10F1B0, bin]
   - [0x1164B8, ci8,     peach_letter, 150, 105]
   - [0x11A23E, bin]
   - [0x11A240, palette, peach_letter]
