+++ ver/us/undefined_syms.txt
@@ -39,68 +39,14 @@ func_80246500_806D80 = 0x80246500;
 func_80242040_8B20B0 = 0x80242040;
 EnterWalk = 0x80285960;
 ExitWalk = 0x80285CF4;
-func_802440E8_804968 = 0x802440E8;
 spiked_goomba = 0x8021B0AC;
 paragoomba = 0x8021CD00;
-StartRumbleWithParams = 0x80267D9C;
 rdpstateinit_dl = 0x80093bb8;
 LeoDiskHandle = 0x800B1B08;
 CartRomHandle = 0x800B4518;
-D_802BFEFC = 0x802BFEFC;
-D_802E9170 = 0x802E9170;
-D_802E91F0 = 0x802E91F0;
-D_802E9270 = 0x802E9270;
-D_802E92B0 = 0x802E92B0;
-D_802E98E0 = 0x802E98E0;
-D_802E9904 = 0x802E9904;
-D_802E99DC = 0x802E99DC;
 nusched = 0x800DA440;
-D_80283770 = 0x80283770;
-D_80291FA8 = 0x80291FA8;
-D_80291FD0 = 0x80291FD0;
-D_80291FF8 = 0x80291FF8;
-D_80292020 = 0x80292020;
-D_80292048 = 0x80292048;
-D_80292070 = 0x80292070;
-D_80292098 = 0x80292098;
-D_802920C0 = 0x802920C0;
-D_802920E8 = 0x802920E8;
-D_80292110 = 0x80292110;
 D_8028358C = 0x8028358C;
 D_800F7BE8 = 0x800F7BE8;
 D_800F7BBC = 0x800F7BBC;
 D_800F7B94 = 0x800F7B94;
 D_802C05CC = 0x802C05CC;
-D_80107CA8 = 0x80107CA8;
-D_80107CF8 = 0x80107CF8;
-D_80107D48 = 0x80107D48;
-D_80107D98 = 0x80107D98;
-D_80107DE8 = 0x80107DE8;
-D_80107CA8 = 0x80107CA8;
-D_80107E88 = 0x80107E88;
-D_80107ED8 = 0x80107ED8;
-D_80107F28 = 0x80107F28;
-D_80107E38 = 0x80107E38;
-D_80107FC8 = 0x80107FC8;
-D_80107CD0 = 0x80107CD0;
-D_80107D20 = 0x80107D20;
-D_80107D70 = 0x80107D70;
-D_80107DC0 = 0x80107DC0;
-D_80107E10 = 0x80107E10;
-D_80107CD0 = 0x80107CD0;
-D_80107EB0 = 0x80107EB0;
-D_80107F00 = 0x80107F00;
-D_80107F50 = 0x80107F50;
-D_80107E60 = 0x80107E60;
-D_80107FF0 = 0x80107FF0;
-D_801080B8 = 0x801080B8;
-D_801080E0 = 0x801080E0;
-D_80108108 = 0x80108108;
-D_80108130 = 0x80108130;
-D_80108158 = 0x80108158;
-D_80108180 = 0x80108180;
-D_801081A8 = 0x801081A8;
-D_801081D0 = 0x801081D0;
-D_801081F8 = 0x801081F8;
-D_80108220 = 0x80108220;
-D_80097D30 = 0x80097D30;
