+++ ver/us/undefined_syms_auto.txt
@@ -3759,7 +3759,6 @@ D_802E9CF8 = 0x802E9CF8;
 D_802E9D1C = 0x802E9D1C;
 D_802E9D40 = 0x802E9D40;
 D_802E9D64 = 0x802E9D64;
-D_802EB380 = 0x802EB380;
 D_802EB390 = 0x802EB390;
 D_802EB394 = 0x802EB394;
 D_802EB398 = 0x802EB398;
